dont retry when it

+++ ci/build/make.sh
@@ -31,7 +31,7 @@ build
 # clean cache and retry
 cached_build_ret=$?
 set -e
-if [ $cached_build_ret -ne 0 ]; then
+if [ $cached_build_ret -ne 0 ] && [[ ! -t 1 ]]; then
     echo "retry after cleaning build dir"
     docker run --rm -v $src_dir:/oneflow-src busybox rm -rf /oneflow-src/manylinux2014-build-cache
     build
